LayoutTests/js/dopey-rope-with-16-bit-propertyname-expected.txt

Test that a 16-bit AtomString containing only 8-bit characters doesn't confuse the JIT into thinking it's an 8-bit AtomString.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS globalRope is 'foo.zest'
PASS successfullyParsed is true

TEST COMPLETE
